Revert "Merge branch 'master' into thesis/2021/multiuser-collab"
This reverts commit 5f07cdbe2e2ccfd3620798f1720e1ba92254018a, reversing
changes made to fc8b7e58d3bf810d816b1b9f6416cbc9c4e04603.

+++ modules/base/rendering/renderableprism.cpp
@@ -215,10 +215,7 @@ void RenderablePrism::updateVertexData() {
     for (int i = 0; i < 2; ++i) {
         float h = i * _length; // z value, 0 to _length
 
-        for (int j = 0, k = 0;
-             j < _nShapeSegments && k < static_cast<int>(unitVertices.size());
-             ++j, k += 2)
-        {
+        for (int j = 0, k = 0; j < _nShapeSegments && k < unitVertices.size(); ++j, k += 2) {
             float ux = unitVertices[k];
             float uy = unitVertices[k + 1];
 
@@ -242,10 +239,7 @@ void RenderablePrism::updateVertexData() {
         _vertexArray.push_back(_length);
     }
     else {
-        for (int j = 0, k = 0;
-             j < _nLines && k < static_cast<int>(unitVerticesLines.size());
-             ++j, k += 2)
-        {
+        for (int j = 0, k = 0; j < _nLines && k < unitVerticesLines.size(); ++j, k += 2) {
             float ux = unitVerticesLines[k];
             float uy = unitVerticesLines[k + 1];
 
@@ -274,8 +268,8 @@ void RenderablePrism::updateVertexData() {
     _indexArray.push_back(255);
 
     // Indices for Top shape
-    for (int i = _nShapeSegments; i < 2 * _nShapeSegments; ++i) {
-        _indexArray.push_back(static_cast<uint8_t>(i));
+    for (uint8_t i = _nShapeSegments; i < 2 * _nShapeSegments; ++i) {
+        _indexArray.push_back(i);
     }
 
     // Indices for connecting lines
@@ -283,8 +277,8 @@ void RenderablePrism::updateVertexData() {
         // Reset
         _indexArray.push_back(255);
 
-        _indexArray.push_back(static_cast<uint8_t>(2 * _nShapeSegments + k));
-        _indexArray.push_back(static_cast<uint8_t>(2 * _nShapeSegments + k + 1));
+        _indexArray.push_back(2 * _nShapeSegments + k);
+        _indexArray.push_back(2 * _nShapeSegments + k + 1);
     }
 }
